+++ spec/interjectable_spec.rb
@@ -6,6 +6,14 @@ describe Interjectable do
   shared_examples_for "an interjectable class" do
     let(:instance) { klass.new }
 
+    def expect_injected_methods(injecting_class, class_instance, *getter_methods)
+      setter_methods = getter_methods.map { |m| :"#{m}=" }
+      expect(injecting_class.injected_methods)
+        .to match_array(getter_methods + setter_methods + [Interjectable::INJECTED_METHODS])
+      expect(injecting_class.injected_methods)
+        .to match_array(class_instance.injected_methods)
+    end
+
     describe "#inject" do
       before do
         klass.inject(:some_dependency) { :service }
@@ -18,12 +26,16 @@ describe Interjectable do
       it "adds an instance method getter and setter" do
         instance.some_dependency = 'aaa'
         expect(instance.some_dependency).to eq('aaa')
+
+        expect_injected_methods(klass, instance, :some_dependency)
       end
 
       it "lazy-loads the default block" do
         expect(instance.instance_variable_get("@some_dependency")).to be_nil
         expect(instance.some_dependency).to eq(:service)
         expect(instance.instance_variable_get("@some_dependency")).not_to be_nil
+
+        expect_injected_methods(klass, instance, :some_dependency)
       end
 
       it "allows transitive dependencies (via instance_eval)" do
@@ -31,6 +43,7 @@ describe Interjectable do
         klass.inject(:second_dependency) { :value }
 
         expect(instance.first_dependency).to eq(:value)
+        expect_injected_methods(klass, instance, :some_dependency, :first_dependency, :second_dependency)
       end
 
       it "calls dependency block once, even with a falsy value" do
@@ -39,6 +52,8 @@ describe Interjectable do
 
         2.times { expect(instance.some_falsy_dependency).to be_nil }
         expect(count).to eq(1)
+
+        expect_injected_methods(klass, instance, :some_dependency, :some_falsy_dependency)
       end
 
       it "errors when injecting the same dependency multiple times" do
@@ -60,6 +75,8 @@ describe Interjectable do
         klass.inject(:good_dependency) { foo }
         instance.foo = 2
         expect(instance.good_dependency).to eq(2)
+
+        expect_injected_methods(klass, instance, :some_dependency, :good_dependency)
       end
 
       context "with a dependency on another class" do
@@ -73,6 +90,7 @@ describe Interjectable do
           instance.some_other_class = :fake_other_class
 
           expect(instance.some_other_class).to eq(:fake_other_class)
+          expect_injected_methods(klass, instance, :some_dependency, :some_other_class)
         end
       end
 
@@ -81,8 +99,20 @@ describe Interjectable do
         let(:subclass_instance) { subclass.new }
 
         it "does not error if the method exists on the superclass" do
-          subclass.inject(:dependency) { :some_other_value }
-          expect(subclass_instance.dependency).to eq(:some_other_value)
+          subclass.inject(:some_dependency) { :some_other_value }
+          expect(subclass_instance.some_dependency).to eq(:some_other_value)
+
+          expect_injected_methods(klass, instance, :some_dependency)
+          expect_injected_methods(subclass, subclass_instance, :some_dependency)
+        end
+
+        it "allows injection on the subclass without injecting on the superclass" do
+          subclass.inject(:subclass_dependency) { :brand_new_value }
+          expect(subclass_instance.subclass_dependency).to eq(:brand_new_value)
+          expect { instance.subclass_dependency }.to raise_error(NoMethodError)
+
+          expect_injected_methods(klass, instance, :some_dependency)
+          expect_injected_methods(subclass, subclass_instance, :some_dependency, :subclass_dependency)
         end
       end
     end
@@ -101,17 +131,26 @@ describe Interjectable do
       it "adds an instance method and setter" do
         instance.static_dependency = 'aaa'
         expect(instance.static_dependency).to eq('aaa')
+
+        expect_injected_methods(klass, instance, :static_dependency)
+        expect_injected_methods(klass, other_instance, :static_dependency)
       end
 
       it "adds a class method and setter" do
         klass.static_dependency = 'aaa'
         expect(klass.static_dependency).to eq('aaa')
+
+        expect_injected_methods(klass, instance, :static_dependency)
+        expect_injected_methods(klass, other_instance, :static_dependency)
       end
 
       it "shares a value across all instances of a class" do
         instance.static_dependency = 'bbb'
         expect(other_instance.static_dependency).to eq('bbb')
         expect(klass.static_dependency).to eq('bbb')
+
+        expect_injected_methods(klass, instance, :static_dependency)
+        expect_injected_methods(klass, other_instance, :static_dependency)
       end
 
       it "calls its dependency block once across all instances" do
@@ -123,6 +162,9 @@ describe Interjectable do
         expect(klass.falsy_static_dependency).to be_nil
 
         expect(count).to eq(1)
+
+        expect_injected_methods(klass, instance, :static_dependency, :falsy_static_dependency)
+        expect_injected_methods(klass, other_instance, :static_dependency, :falsy_static_dependency)
       end
 
       it "errors when inject_static-ing a dependency multiple times" do
@@ -163,23 +205,49 @@ describe Interjectable do
           instance.static_dependency = 'ccc'
           expect(subclass_instance.static_dependency).to eq('ccc')
           expect(subclass.static_dependency).to eq('ccc')
+
+          expect_injected_methods(klass, instance, :static_dependency)
+          expect_injected_methods(klass, other_instance, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, :static_dependency)
         end
 
         it "does not error if the method exists on the super klass" do
           subclass.inject_static(:static_dependency) { :some_other_value }
           expect(subclass_instance.static_dependency).to eq(:some_other_value)
           expect(subclass.static_dependency).to eq(:some_other_value)
+
+          expect_injected_methods(klass, instance, :static_dependency)
+          expect_injected_methods(klass, other_instance, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, :static_dependency)
         end
 
         it "does not error when lazily setting the dep from a subclass first" do
           expect(subclass.static_dependency).to eq(:some_value)
           expect(klass.static_dependency).to eq(:some_value)
           expect(subclass.static_dependency).to eq(:some_value)
+
+          expect_injected_methods(klass, instance, :static_dependency)
+          expect_injected_methods(klass, other_instance, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, :static_dependency)
         end
 
         it "only defines the class variable on the injecting class" do
           expect(subclass.static_dependency).to eq(:some_value)
           expect(klass.class_variable_get(:@@static_dependency)).to eq(:some_value)
+
+          expect_injected_methods(klass, instance, :static_dependency)
+          expect_injected_methods(klass, other_instance, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, :static_dependency)
+        end
+
+        it "allows injection on the subclass without injecting on the superclass" do
+          subclass.inject_static(:static_subclass_dependency) { :brand_new_value }
+          expect(subclass_instance.static_subclass_dependency).to eq(:brand_new_value)
+          expect { klass.static_subclass_dependency }.to raise_error(NoMethodError)
+          expect { instance.static_subclass_dependency }.to raise_error(NoMethodError)
+
+          expect_injected_methods(klass, instance, :static_dependency)
+          expect_injected_methods(subclass, subclass_instance, :static_dependency, :static_subclass_dependency)
         end
       end
     end